LPD-91262 Move the functionality to prevent navigation in an iframe to layout-js-components-web
