+++ modules/apps/layout/layout-js-components-web/src/main/resources/META-INF/resources/js/index.ts
@@ -51,3 +51,4 @@ export {default as convertRGBtoHex} from './utils/convertRGBtoHex';
 export {default as isCtrlOrMeta} from './utils/isCtrlOrMeta';
 export {default as isNullOrUndefined} from './utils/isNullOrUndefined';
 export {default as isValidStyleValue} from './utils/isValidStyleValue';
+export {default as preventIframeNavigation} from './utils/preventIframeNavigation';
